no pointers; working solver initialization with plant --->
Not using pointers for Geometry and Assimilator in Plant,
to avoid dealing with Rule of 5. Copy constructor is
crucial because solver relies on it to insert copies of
boundary cohort.

made proper state and rate structs in plant for better
organization.

Working initialization for FMU solver. Environment needs
to be set BEFORE initializing solver, as establishmentProb()
needs env.